+++ src/airunner/widgets/canvas/standard_image_widget.py
@@ -7,7 +7,7 @@ from PyQt6.QtGui import QImage
 from PIL import Image
 
 from airunner.enums import SignalCode, ServiceCode
-from airunner.widgets.canvas_plus.templates.standard_image_widget_ui import Ui_standard_image_widget
+from airunner.widgets.canvas.templates.standard_image_widget_ui import Ui_standard_image_widget
 from airunner.utils import load_metadata_from_image, prepare_metadata
 from airunner.widgets.slider.slider_widget import SliderWidget
 from airunner.widgets.base_widget import BaseWidget
