+++ src/airunner/widgets/canvas/templates/brushes_container.ui
