+++ src/airunner/widgets/canvas/templates/standard_image_widget.ui
