+++ src/airunner/widgets/canvas/templates/standard_video_widget.ui
